feat: MmtileWriter + BakeBlackspots - save modified tiles to disk
- Add MmtileWriter.cs: writes NavMeshData back to .mmtile format (exact mirror of MmtileLoader)
- Fix DT_POLYREF64: NavLink.Ref is ulong (8 bytes), dtLink = 16 bytes not 12
- Add MeshBaker.BakeBlackspots(): marks polys in blackspot cylinders as AREA_UNWALKABLE(63)
- Add menu: Mesh > Bake Blackspots into Tile + Save Modified Tile (.mmtile)
- Verified round-trip: 2/2 tiles PERFECT MATCH (size formula validated)

+++ Core/MmtileLoader.cs
@@ -200,7 +200,7 @@ namespace MeshViewer3D.Core
         }
 
         /// <summary>
-        /// Lit les liens entre polygones (8 bytes par link)
+        /// Lit les liens entre polygones (16 bytes par link avec DT_POLYREF64)
         /// </summary>
         private static NavLink[] ReadLinks(BinaryReader br, int count)
         {
@@ -209,7 +209,7 @@ namespace MeshViewer3D.Core
             {
                 links[i] = new NavLink
                 {
-                    Ref = br.ReadUInt32(),
+                    Ref = br.ReadUInt64(),
                     Next = br.ReadUInt32(),
                     Edge = br.ReadByte(),
                     Side = br.ReadByte(),
